topcat: fixed export bug in Stats window
The Sum column in the StatisticsWindow's TableModel was incorrect:
it declared a contentClass of Double, but for boolean data table
columns it was yielding a Long.  That wasn't a problem for the
JTable display, but it led to a class incompatibility when
attempting to export the stats (save to disk or import as new table),
manifesting as a stacktrace preventing the export action.

Thanks to Henrik Rydberg for reporting this.

+++ topcat/src/main/uk/ac/starlink/topcat/StatsWindow.java
@@ -440,7 +440,7 @@ public class StatsWindow extends AuxWindow {
                     return new Double( lastCalc_.sums[ jcol ] );
                 }
                 else if ( lastCalc_.isBoolean[ jcol ] ) {
-                    return new Long( lastCalc_.ntrues[ jcol ] );
+                    return new Double( lastCalc_.ntrues[ jcol ] );
                 }
                 else {
                     return null;
